+++ gluten-core/src/main/scala/io/glutenproject/extension/columnar/CollapseProjectExecTransformer.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.glutenproject.extension
+package io.glutenproject.extension.columnar
 
 import io.glutenproject.GlutenConfig
 import io.glutenproject.execution.ProjectExecTransformer
