+++ gluten-ut/spark33/src/test/scala/org/apache/spark/sql/execution/FallbackStrategiesSuite.scala
@@ -18,8 +18,8 @@ package org.apache.spark.sql.execution
 
 import io.glutenproject.backendsapi.BackendsApiManager
 import io.glutenproject.execution.BasicScanExecTransformer
-import io.glutenproject.extension.{ColumnarOverrideRules, GlutenPlan, InsertTransitions}
-import io.glutenproject.extension.columnar.{FallbackEmptySchemaRelation, TRANSFORM_UNSUPPORTED, TransformHints}
+import io.glutenproject.extension.{ColumnarOverrideRules, GlutenPlan}
+import io.glutenproject.extension.columnar.{FallbackEmptySchemaRelation, InsertTransitions, TRANSFORM_UNSUPPORTED, TransformHints}
 import io.glutenproject.utils.QueryPlanSelector
 
 import org.apache.spark.rdd.RDD
